requirements.txt
openai-agents
mcp
httpx
python-dotenv
